Enforce (dll|exe)$ for IsBounceExecutable
Configuration files in a project included in a build will be copied to
the bin directory with a `.dll.config` extension.

These files are trying to be `Assembly.Load`ed causing the BounceRunner
task to error out.

In this fix `IsBounceExecutable` enforces that the files must end will
dll or exe. This matches the existing readme.